feat: Update chatgpt url
