+++ lib/pages/chatbot_page.dart
@@ -1,21 +1,211 @@
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
-import 'package:nihongo/chatbot/microphone.dart';
+import 'package:flutter_spinkit/flutter_spinkit.dart'; // For 3-dot animation
+import 'package:http/http.dart' as http;
+import 'package:nihongo/services/audio_record.dart';
+import 'package:nihongo/services/audio_play.dart';
+import 'package:nihongo/services/speech_to_text.dart';
 import 'package:nihongo/services/text_to_speech.dart';
 
-// Create a TextToSpeech instance
-final textToSpeech = TextToSpeech();
-const microphoneButton = MicrophoneButton();
-
-class ChatbotPage extends StatelessWidget {
+class ChatbotPage extends StatefulWidget {
   const ChatbotPage({super.key});
+
+  @override
+  _ChatbotPageState createState() => _ChatbotPageState();
+}
+
+class _ChatbotPageState extends State<ChatbotPage> {
+  final AudioRecord audioRecord = AudioRecord();
+  final AudioPlay audioPlay = AudioPlay();
+  final SpeechToText speechToText = SpeechToText();
+  final TextToSpeech textToSpeech = TextToSpeech();
+  final TextEditingController _textController = TextEditingController();
+  final List<Map<String, dynamic>> _messages = [];
+  bool _isThinking = false;
+  bool _isRecording = false;
+  bool _isPlaying = false;
+
+  final String serverUrl = "https://nihongobenkyou.online/api/chat/";
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        backgroundColor: Colors.grey[200],
-        appBar: const ChatAppBar(),
-        body: const ChatBody(),
-        bottomNavigationBar: const ChatBottomNavigationBar());
+      backgroundColor: Colors.grey[200],
+      appBar: const ChatAppBar(),
+      body: Column(
+        children: [
+          Expanded(child: _buildChatBody()),
+          _buildBottomNavigationBar(),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildChatBody() {
+    return ListView.builder(
+      padding: const EdgeInsets.all(16.0),
+      itemCount: _messages.length + (_isThinking ? 1 : 0),
+      itemBuilder: (context, index) {
+        if (index == _messages.length) {
+          // Show "thinking" animation
+          return Align(
+            alignment: Alignment.centerLeft,
+            child: Container(
+              margin: const EdgeInsets.symmetric(vertical: 8.0),
+              padding: const EdgeInsets.all(12.0),
+              decoration: BoxDecoration(
+                color: Colors.white,
+                borderRadius: BorderRadius.circular(12),
+              ),
+              child: const SpinKitThreeBounce(
+                color: Colors.grey,
+                size: 20.0,
+              ),
+            ),
+          );
+        }
+        final message = _messages[index];
+        final isUserMessage = message['isUser'];
+        return Align(
+          alignment:
+              isUserMessage ? Alignment.centerRight : Alignment.centerLeft,
+          child: Container(
+            margin: const EdgeInsets.symmetric(vertical: 8.0),
+            padding: const EdgeInsets.all(12.0),
+            decoration: BoxDecoration(
+              color: isUserMessage ? Colors.blue[100] : Colors.white,
+              borderRadius: BorderRadius.circular(12),
+            ),
+            child: Text(message['text']),
+          ),
+        );
+      },
+    );
+  }
+
+  Widget _buildBottomNavigationBar() {
+    return Padding(
+      padding: const EdgeInsets.all(8.0),
+      child: Row(
+        children: [
+          Expanded(
+            child: TextField(
+              controller: _textController,
+              decoration: InputDecoration(
+                hintText: "Type a message...",
+                filled: true,
+                fillColor: Colors.white,
+                border: OutlineInputBorder(
+                  borderSide: BorderSide.none,
+                  borderRadius: BorderRadius.circular(20),
+                ),
+                contentPadding:
+                    const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
+              ),
+            ),
+          ),
+          const SizedBox(width: 10),
+          IconButton(
+            icon: Icon(Icons.mic,
+                color: _isRecording ? Colors.red : Colors.purple),
+            onPressed: () {
+              if (_isRecording) {
+                audioRecord.stopRecording();
+                speechToText.speechToText().then((newText) {
+                  setState(() {
+                    _textController.text = newText;
+                  });
+                });
+                // speechToText.speechToText().then((value) {
+                //   setState(() {
+                //     _textController.text = value;
+                //   });
+                // });
+                // audioPlay.startPlaying();
+                setState(() {
+                  _isRecording = false;
+                });
+              } else {
+                audioRecord.startRecording();
+                setState(() {
+                  _isRecording = true;
+                });
+              }
+            },
+          ),
+          IconButton(
+            icon: const Icon(Icons.send, color: Colors.purple),
+            onPressed: _sendMessage,
+          ),
+        ],
+      ),
+    );
+  }
+
+  Future<void> _stopRecordAudio() async {
+    audioRecord.stopRecording();
+    speechToText.speechToText().then((value) {
+      setState(() {
+        _textController.text = value;
+      });
+    });
+    setState(() {
+      _isRecording = false;
+    });
+  }
+
+  // Future<File> _simulateAudioRecording() async {
+  //   // Simulate recording a .wav file
+  //   return File('path/to/audio.wav'); // Replace with actual recorded file
+  // }
+
+  // Future<String> _sendAudioToServer(File audioFile) async {
+  //   final request = http.MultipartRequest("POST", Uri.parse(serverUrl))
+  //     ..files.add(await http.MultipartFile.fromPath('file', audioFile.path));
+
+  //   final response = await request.send();
+  //   if (response.statusCode == 200) {
+  //     final responseData = await response.stream.bytesToString();
+  //     return jsonDecode(responseData)['text'];
+  //   } else {
+  //     return "Error processing audio.";
+  //   }
+  // }
+
+  Future<void> _sendMessage() async {
+    final userText = _textController.text;
+    if (userText.isEmpty) return;
+
+    setState(() {
+      _messages.add({'text': userText, 'isUser': true});
+      _textController.clear();
+      _isThinking = true;
+    });
+
+    try {
+      final response = await http.post(
+        Uri.parse(serverUrl),
+        headers: {"Content-Type": "application/json"},
+        body: jsonEncode({"message": userText}),
+      );
+
+      if (response.statusCode == 200) {
+        final responseText = jsonDecode(response.body)['response'];
+        setState(() {
+          _isThinking = false;
+          _messages.add({'text': responseText, 'isUser': false});
+        });
+      } else {
+        throw Exception("Failed to send message");
+      }
+    } catch (e) {
+      setState(() {
+        _isThinking = false;
+        _messages
+            .add({'text': "Error processing message: $e", 'isUser': false});
+      });
+    }
   }
 }
 
@@ -29,21 +219,11 @@ class ChatAppBar extends StatelessWidget implements PreferredSizeWidget {
       backgroundColor: const Color(0xFF8980F0),
       title: const Row(
         children: [
-          Icon(Icons.rocket, color: Colors.white, size: 28),
+          Icon(Icons.chat, color: Colors.white, size: 28),
           SizedBox(width: 10),
           Text(
             "ChatGPT",
-            style: TextStyle(
-              fontSize: 24,
-              fontFamily: 'Noto Sans',
-              fontWeight: FontWeight.w700,
-              color: Colors.white,
-            ),
-          ),
-          Spacer(),
-          Text(
-            "Online",
-            style: TextStyle(color: Colors.white, fontSize: 18),
+            style: TextStyle(fontSize: 24, fontWeight: FontWeight.w700),
           ),
         ],
       ),
@@ -52,160 +232,4 @@ class ChatAppBar extends StatelessWidget implements PreferredSizeWidget {
 
   @override
   Size get preferredSize => const Size.fromHeight(kToolbarHeight);
-}
-
-class ChatBody extends StatefulWidget {
-  const ChatBody({super.key});
-
-  @override
-  _ChatBodyState createState() => _ChatBodyState();
-}
-
-class TextFieldFocus extends StatefulWidget {
-  final TextEditingController controller;
-
-  const TextFieldFocus({super.key, required this.controller});
-
-  @override
-  _TextFieldFocusState createState() => _TextFieldFocusState();
-}
-
-class _TextFieldFocusState extends State<TextFieldFocus> {
-  final FocusNode _focus = FocusNode();
-
-  @override
-  void initState() {
-    super.initState();
-    _focus.addListener(_onFocusAndRecordingChange);
-  }
-
-  @override
-  void dispose() {
-    _focus.removeListener(_onFocusAndRecordingChange);
-    _focus.dispose();
-    super.dispose();
-  }
-
-  void _onFocusAndRecordingChange() {
-    _onFocusChange();
-    _onRecording();
-  }
-
-  void _onRecording() {
-    debugPrint("debug: Recording ${microphoneButton.isRecording}");
-  }
-
-  void _onFocusChange() {
-    debugPrint("Focus: ${_focus.hasFocus.toString()}");
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return TextField(
-      controller: widget.controller,
-      focusNode: _focus,
-      decoration: InputDecoration(
-        hintText: "Nhập tin nhắn...",
-        filled: true,
-        fillColor: Colors.white,
-        border: OutlineInputBorder(
-          borderSide: BorderSide.none,
-          borderRadius: BorderRadius.circular(20),
-        ),
-        contentPadding:
-            const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
-      ),
-    );
-  }
-}
-
-class _ChatBodyState extends State<ChatBody> {
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(16.0),
-      child: SingleChildScrollView(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            _buildOptionTile(
-              icon: Icons.translate,
-              title: "Dịch",
-              subtitle: "Tôi có thể giúp bạn dịch",
-            ),
-            const SizedBox(height: 20),
-            _buildOptionTile(
-              icon: Icons.edit,
-              title: "Viết câu hỏi",
-              subtitle: "Hãy hỏi tôi điều gì đó",
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildOptionTile(
-      {required IconData icon,
-      required String title,
-      required String subtitle}) {
-    return Container(
-      padding: const EdgeInsets.all(16),
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(12),
-      ),
-      child: Row(
-        children: [
-          Icon(icon, size: 30, color: Colors.grey[700]),
-          const SizedBox(width: 16),
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                title,
-                style:
-                    const TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
-              ),
-              const SizedBox(height: 4),
-              Text(
-                subtitle,
-                style: TextStyle(fontSize: 14, color: Colors.grey[600]),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class ChatBottomNavigationBar extends StatelessWidget {
-  const ChatBottomNavigationBar({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    TextEditingController controller = TextEditingController();
-
-    return Padding(
-      padding: const EdgeInsets.all(8.0),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Expanded(
-            child: TextFieldFocus(
-              controller: controller,
-            ),
-          ),
-          const SizedBox(width: 10),
-          microphoneButton,
-          IconButton(
-            onPressed: () {},
-            icon: const Icon(Icons.photo_camera),
-            color: Colors.purple,
-          ),
-        ],
-      ),
-    );
-  }
 }
